@@ -1,5 +1,5 @@
 {
-  "name": "pocketbase-sveltekit-auth",
+  "name": "super-garde",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -12,6 +12,12 @@
     "format": "prettier --plugin-search-dir . --write ."
   },
   "devDependencies": {
+    "@event-calendar/core": "^2.4.1",
+    "@event-calendar/day-grid": "^2.4.1",
+    "@event-calendar/interaction": "^2.4.1",
+    "@event-calendar/list": "^2.4.1",
+    "@event-calendar/resource-time-grid": "^2.4.1",
+    "@event-calendar/time-grid": "^2.4.1",
     "@sveltejs/adapter-auto": "^2.1.0",
     "@sveltejs/kit": "^1.25.1",
     "@tailwindcss/forms": "^0.5.6",
@@ -19,15 +25,15 @@
     "@typescript-eslint/eslint-plugin": "^6.7.4",
     "@typescript-eslint/parser": "^6.7.4",
     "autoprefixer": "^10.4.16",
-    "daisyui": "^3.8.3",
-    "eslint": "^8.50.0",
+    "daisyui": "^4.4.2",
+    "eslint": "^8.54.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-svelte3": "^4.0.0",
+    "eslint-plugin-svelte": "^2.35.1",
     "postcss": "^8.4.31",
     "postcss-load-config": "^4.0.1",
     "prettier": "^3.0.3",
     "prettier-plugin-svelte": "^3.0.3",
-    "svelte": "^4.2.1",
+    "svelte": "^4.2.7",
     "svelte-check": "^3.5.2",
     "svelte-preprocess": "^5.0.4",
     "tailwindcss": "^3.3.3",
