@@ -1,4 +1,8 @@
-import type { Admin, Record } from 'pocketbase'
-import { writable } from 'svelte/store'
+import type { Admin, Record } from 'pocketbase';
+import { writable } from 'svelte/store';
 
-export const currentUser = writable<Record | Admin | null>()
+export interface UserRecord extends Record {
+  role: 'student' | 'admin' | 'god';
+}
+
+export const currentUser = writable<UserRecord | Admin | null>()
