@@ -1,4 +1,6 @@
-<h1>Welcome to SvelteKit</h1>
+<script lang="ts">
+</script>
+
 <p>
-  Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation
+  Salut, moi c'est Elfo!
 </p>
